refactor: 60-move router to app folder
